refactor: 차트 컴포넌트 네이밍 수정 (#151)
* refactor: 차트 컴포넌트 네이밍 Radar 제거

* refactor: 차트 컴포넌트 리네이밍에 맞춰 Charts 탭 갱신
