+++ src/components/Charts/Charts.tsx
@@ -4,8 +4,8 @@ import { useState } from 'react';
 import { LineChart, Radar, Weight } from 'lucide-react';
 
 import RecordChart from './components/RecordChart/RecordChart';
-import StrengthRadarChart from './components/StrengthRadarChart/StrengthRadarChart';
-import BodyweightRadarChart from './components/BodyweightRadarChart/BodyweightRadarChart';
+import StrengthChart from './components/StrengthChart/StrengthChart';
+import BodyweightChart from './components/BodyweightChart/BodyweightChart';
 
 import styles from './Charts.module.scss';
 
@@ -13,11 +13,11 @@ type ChartsProps = {
   userId?: string;
 };
 
-type ChartView = 'record' | 'radar' | 'bodyweight';
+type ChartView = 'record' | 'strength' | 'bodyweight';
 
 const TABS: { key: ChartView; label: string; icon: React.ReactNode }[] = [
   { key: 'record', label: '성장 곡선', icon: <LineChart size={20} /> },
-  { key: 'radar', label: '밸런스 분석', icon: <Radar size={20} /> },
+  { key: 'strength', label: '밸런스 분석', icon: <Radar size={20} /> },
   { key: 'bodyweight', label: '체중 대비', icon: <Weight size={20} /> },
 ];
 
@@ -44,11 +44,11 @@ export default function Charts({ userId }: ChartsProps) {
       {activeView === 'record' && (
         <RecordChart userId={userId} tabButtons={tabButtons} />
       )}
-      {activeView === 'radar' && (
-        <StrengthRadarChart userId={userId} tabButtons={tabButtons} />
+      {activeView === 'strength' && (
+        <StrengthChart userId={userId} tabButtons={tabButtons} />
       )}
       {activeView === 'bodyweight' && (
-        <BodyweightRadarChart userId={userId} tabButtons={tabButtons} />
+        <BodyweightChart userId={userId} tabButtons={tabButtons} />
       )}
     </div>
   );
