+++ src/components/Charts/components/BodyweightChart/BodyweightChart.module.scss
